fix bug and updated code

+++ src/main/java/com/huseynov/restaurant/cart/CartServiceImpl.java
@@ -3,6 +3,7 @@ package com.huseynov.restaurant.cart;
 import com.huseynov.restaurant.customer.Customer;
 import com.huseynov.restaurant.customer.CustomerService;
 import com.huseynov.restaurant.shared.exception.CustomNotFoundException;
+import com.huseynov.restaurant.shared.exception.InvalidRequestException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -28,10 +29,15 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public Cart getCart() {
-        Long id = cartRepository
-                .findCartIdByCustomerId(customer.getId())
-                .orElseThrow(() -> new CustomNotFoundException("Cart not found for customer,  id: " + customer.getId()));
-        log.info("CartServiceImpl::getCart , id: {}", id);
+        log.info("CartServiceImpl::getCart");
+
+        Long id = getCartId();
+        if (id == null) {
+            log.error("CartServiceImpl::getCart , cart not found for customer, id: {}", customer.getId());
+            throw new CustomNotFoundException("Cart not found for customer, id: " + customer.getId());
+        }
+
+        log.info("CartServiceImpl::getCartById , id: {}", id);
         return cartRepository
                 .findCartByCartId(id)
                 .orElseThrow(() -> new CustomNotFoundException("Cart not found, id: " + id));
@@ -45,9 +51,14 @@ public class CartServiceImpl implements CartService {
 
     @Override
     public void deleteCart() {
-        Long id = cartRepository
-                .findCartIdByCustomerId(customer.getId())
-                .orElseThrow(() -> new CustomNotFoundException("Cart not found for customer, id: " + customer.getId()));
+        log.info("CartServiceImpl::deleteCart");
+
+        Long id = getCartId();
+        if (id == null) {
+            log.error("CartServiceImpl::deleteCart , cart not found for customer, id: {}", customer.getId());
+            throw new CustomNotFoundException("Cart not found for customer,  id: " + customer.getId());
+        }
+
         log.info("CartServiceImpl::deleteCartById , id: {}", id);
         cartRepository.deleteById(id);
     }
@@ -61,16 +72,6 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public boolean isCustomerRequestValidForCart() {
-        Long cartId = cartRepository
-                .findCartIdByCustomerId(customer.getId())
-                .orElseThrow(() -> new CustomNotFoundException("Cart not found for customer, id: " + customer.getId()));
-
-        log.info("CartServiceImpl::isCustomerRequestValidForCart , cartId: {}", cartId);
-        return cartRepository.existsCartByCustomerIdAndCartId(customer.getId(), cartId);
-    }
-
-    @Override
     public Long getCartId() {
         log.info("CartServiceImpl::getCartId");
         return cartRepository
@@ -78,5 +79,33 @@ public class CartServiceImpl implements CartService {
                 .orElse(null);
     }
 
+    /**
+     * Checks if the customer request is valid for the cart.
+     *
+     * @return true if the customer request is valid for the cart, false otherwise
+     * @throws CustomNotFoundException if the cart is not found for the customer
+     */
+    @Override
+    public boolean isCustomerRequestValidForCart() {
+        log.info("CartServiceImpl::isCustomerRequestInvalidForCart");
+
+        Long id = getCartId();
+        if (id == null) {
+            log.error("CartServiceImpl::isCustomerRequestInvalidForCart , cart not found for customer, id: {}", customer.getId());
+            throw new CustomNotFoundException("Cart not found for customer,  id: " + customer.getId());
+        }
+
+        log.info("CartServiceImpl::isCustomerRequestValidForCart , cartId: {}", id);
+        return cartRepository.existsCartByCustomerIdAndCartId(customer.getId(), id);
+    }
+
+    // if the customer who sent request is not the owner of the cart(for exists cart)
+    @Override
+    public void checkCustomerForCart() {
+        log.info("CartServiceImpl::checkCustomerForCart");
+        if (getCartId() != null && !isCustomerRequestValidForCart()) {
+            throw new InvalidRequestException("Invalid request for cart ");
+        }
+    }
 
 }
